chore(posts): Move `SearchParams` out of `photo`s since they're shared.
Or at least, they should be.
